+++ src/vsc-leetcode-cli/new_lib/my_plugin_base.ts
@@ -99,6 +99,9 @@ export class MyPluginBase {
   public getUserContestP(username, cb) {
     this.next.getUserContestP(username, cb)
   }
+  public getProblemsTitle(cb) {
+    this.next.getProblemsTitle(cb)
+  }
 }
