+++ src/vsc-leetcode-cli/new_lib/session.ts
@@ -14,6 +14,9 @@ class Session {
       statusCode: -1
     }
   };
+  argv: any = {
+
+  }
   constructor() {
 
   }
